HSEARCH-3810 Move SimpleNameClassFormatter to a more appropriate module
There's nothing specific to the engine in this class.
